bugfix 初期設定部分にバグがあったのを修正しました。

+++ Muremon/Program/Actor/ActorBoss.h
@@ -24,7 +24,7 @@ public:
 	bool		isWin() const;
 
 public:
-	float		mLife;
+	int			mLife;
 	float		mMaxLife;
 	float		mMoveX;
 	float		mMoveY;
